refactor: rename config.ts to types.ts
